Update LDAP integration test to work in admin context
